Minor refactor for prettier
