--- client/pages/_app.tsx
@@ -1,31 +1,0 @@
-import React from "react";
-import App, { Container } from "next/app";
-import Layout from "../components/layout";
-import "../styles/reset.scss";
-
-class MyApp extends App {
-  static async getInitialProps({ Component, ctx }) {
-    let pageProps = {};
-
-    if (Component.getInitialProps) {
-      pageProps = await Component.getInitialProps(ctx);
-    }
-
-    return { pageProps };
-  }
-
-  render() {
-    const {props} = this as any
-    const {Component, pageProps} = props;
-    
-    return (
-      <Container>
-        <Layout>
-          <Component {...pageProps} />
-        </Layout>
-      </Container>
-    );
-  }
-}
-
-export default MyApp;
