+++ dashboard/client/pages/_error.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react'
+
+const Error = () => <>☠️</>
+
+export default Error
