+++ dashboard/client/pages/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from 'next'
+import React from 'react'
+
+const HomePage: NextPage = () => {
+  return <div>Welcome to Next.js!</div>
+}
+
+export default HomePage
